backend/requirements.txt
flask
mysql-connector-python
opencv-python
numpy
ultralytics
